Fix ISession<TAggregate, IRepository<TAggregate>
