+++ Learning.NETCore/AggregateRoot/Learning.AggregateRoot.Domain/ExampleToDelete/Builders/IItemViewMapper.cs
@@ -1,0 +1,10 @@
+﻿using Learning.AggregateRoot.Domain.ExampleToDelete.Aggregates;
+using Learning.AggregateRoot.Domain.ExampleToDelete.Views;
+
+namespace Learning.AggregateRoot.Domain.ExampleToDelete.Builders
+{
+    public interface IItemViewMapper
+    {
+        ItemView Map(Item item);
+    }
+}
