+++ Learning.NETCore/AggregateRoot/Learning.AggregateRoot.Infrastructure/ExampleToRemove/QueryHandlers/ItemQueryHandler.cs
@@ -1,34 +1,40 @@
-﻿using System.Linq;
-using System.Threading;
+﻿using System.Threading;
 using System.Threading.Tasks;
 using Learning.AggregateRoot.Domain.CQRS.Interfaces;
 using Learning.AggregateRoot.Domain.ExampleToDelete.Aggregates;
+using Learning.AggregateRoot.Domain.ExampleToDelete.Builders;
 using Learning.AggregateRoot.Domain.ExampleToDelete.Queries;
+using Learning.AggregateRoot.Domain.ExampleToDelete.Repositories;
 using Learning.AggregateRoot.Domain.ExampleToDelete.Views;
 using MediatR;
 
 namespace Learning.AggregateRoot.Infrastructure.ExampleToRemove.QueryHandlers
 {
-    public class ItemQueryHandler : IRequestHandler<GetItem, ItemView>
+    public class ItemQueryHandler :
+        IRequestHandler<GetItem, ItemView>,
+        IRequestHandler<GetItemByName, ItemView>
     {
-        private readonly ISession<Item> _session;
+        private readonly ISession<Item, IItemRepository> _session;
+        private readonly IItemViewMapper _itemViewMapper;
 
-        public ItemQueryHandler(ISession<Item> session) => _session = session;
-        
+        public ItemQueryHandler(ISession<Item, IItemRepository> session, IItemViewMapper itemViewMapper)
+        {
+            _session = session;
+            _itemViewMapper = itemViewMapper;
+        }
+
         public async Task<ItemView> Handle(GetItem query, CancellationToken cancellationToken)
         {
             var item = _session.GetActive(query.Id, _ => _.Locations);
 
-            return new ItemView
-            {
-                Id = item.Id,
-                Name = item.Name,
-                Locations = item.Locations.Select(l => new ItemLocationView
-                {
-                    Id = l.Id,
-                    Name = l.Name
-                })
-            };
+            return _itemViewMapper.Map(item);
+        }
+
+        public async Task<ItemView> Handle(GetItemByName query, CancellationToken cancellationToken)
+        {
+            var item = _session.Repository.GetByName(query.Name);
+
+            return _itemViewMapper.Map(item);
         }
     }
 }
